requirements.txt
streamlit
langchain
langchain-community
langchainhub
langchain-google-genai
PyMuPDF
faiss-cpu
python-dotenv